app/requirements.txt
flask==3.0.3
pyodbc==5.1.0
requests==2.32.3
azure-storage-blob==12.20.0
